+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add Compact Log Event Format (CLEF) formatter.
+
 ## [0.13.0] - 2025-05-06
 
 ### Added